requirements.txt
asgiref==3.8.1
asttokens==2.4.1
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
Flask==3.0.3
fonttools==4.47.0
idna==3.7
ipykernel==6.28.0
ipython==8.19.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
jplephem==2.22
jupyter_client==8.6.0
jupyter_core==5.7.0
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.3
oauthlib==3.2.2
packaging==23.2
pandas==2.1.4
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
pyzmq==25.1.2
requests==2.32.3
requests-oauthlib==1.3.1
scikit-learn==1.3.2
scipy==1.11.4
sgp4==2.23
six==1.16.0
skyfield==1.49
sqlparse==0.5.1
stack-data==0.6.3
threadpoolctl==3.2.0
tornado==6.4
traitlets==5.14.1
tweepy==4.14.0
typing_extensions==4.12.2
tzdata==2023.4
urllib3==2.2.2
wcwidth==0.2.12
Werkzeug==3.0.3
waitress
